Cleanup work on the originate functions.
Remove all notion of returning the call UUID from the helpers and leave that up to the users. In replacement of that there is a new struct for building each leg of the call allowing more flexibility.
