+++ example/inbound/inbound.go
@@ -31,9 +31,9 @@ func main() {
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
 	defer cancel()
 
-	// Place the call to user 100 and playback an audio file as the bLeg
-	originationUUID, response, err := conn.OriginateCall(ctx, true, "user/100", "&playback(misc/ivr-to_hear_screaming_monkeys.wav)", map[string]string{})
-	fmt.Println("Call Originated: ", originationUUID, response, err)
+	// Place the call in the background(bgapi) to user 100 and playback an audio file as the bLeg and no exported variables
+	response, err := conn.OriginateCall(ctx, true, eslgo.Leg{CallURL: "user/100"}, eslgo.Leg{CallURL: "&playback(misc/ivr-to_hear_screaming_monkeys.wav)"}, map[string]string{})
+	fmt.Println("Call Originated: ", response, err)
 
 	// Close the connection after sleeping for a bit
 	time.Sleep(60 * time.Second)
